+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Toutes les modifications notables de ce projet seront documentées dans ce fichi
 - [#242](https://github.com/Refhi/Weda-Helper/issues/242) - lors de l'utilisation de raccourcis claviers, plusieurs consultations pouvaient être créées en même temps. Ajout d'un délais réfractaire pour éviter cela.
 - [#231](https://github.com/Refhi/Weda-Helper/issues/231) - l'inversion automatique des champs ALD se déclenchait trop facilement
 - [#228](https://github.com/Refhi/Weda-Helper/issues/228) - tentative de fix de la mauvaise selection du règlement "Virement" lors des téléconsultations
-- neutralisation de l'ancienne méthode d'anti-throttle
+- suppression de l'ancienne méthode d'anti-throttle
   
 ## ajouts :
 - ajout d'un lien avec Weda pour la désactivation des options de Weda Helper déjà implémentées dans Weda.